+++ lib/t/test_matpower.m
@@ -32,6 +32,7 @@ tests = {};
 %% MATPOWER base test
 tests{end+1} = 't_test_fcns';
 tests{end+1} = 't_nested_struct_copy';
+tests{end+1} = 't_have_fcn';
 tests{end+1} = 't_feval_w_path';
 tests{end+1} = 't_mpoption';
 tests{end+1} = 't_loadcase';